+++ ClusterSets.py
@@ -5,8 +5,8 @@ Cluster sequences by group
 
 __author__    = 'Christopher Bolen, Jason Vander Heiden'
 __license__   = 'Creative Commons Attribution-NonCommercial-ShareAlike 3.0 Unported'
-__version__   = '0.4.7'
-__date__      = '2015.06.05'
+__version__   = '0.4.8'
+__date__      = '2015.08.22'
 
 # Imports
 import csv, os, sys, tempfile, textwrap, time