+++ tests/test_ConvertHeaders.py
@@ -7,7 +7,7 @@ from __future__ import absolute_import, division, print_function
 __author__    = 'Jason Anthony Vander Heiden'
 __copyright__ = 'Copyright 2014 Kleinstein Lab, Yale University. All rights reserved.'
 __license__   = 'Creative Commons Attribution-NonCommercial-ShareAlike 3.0 Unported'
-__version__   = '0.4.7'
+__version__   = '0.4.8'
 __date__      = '2015.05.09'
 
 # Imports